Fixed validation problems and divided zod schema into different files for better organization.
